requirements.txt
streamlit
python-dotenv
notion-client
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
langchain
langchain-community
langchain-ollama
langchain-huggingface
chromadb
transformers

